requirements.txt
transformers
requests
langchain
streamlit
openai==0.28
python-dotenv